updated more - link to worktree method works now too!

+++ deploy-script.sh
@@ -2,7 +2,13 @@
 # any error, whole script fails
 set -e
 
-# Subtree method here: https://gohugo.io/hosting-and-deplooyment/hosting-on-github/
+if [ "`git status -s`" ]
+then
+    echo "The working directory is dirty. Please commit any pending changes."
+    exit 1;
+fi
+
+# Subtree method here: https://gohugo.io/hosting-and-deployment/hosting-on-github/
 echo "Deleting old publication"
 rm -rf public
 mkdir public
@@ -10,7 +16,7 @@ git worktree prune
 rm -rf .git/worktrees/public/
 
 echo "Checking out deploy branch into public"
-git worktree add -B prod public
+git worktree add -B prod public deploy/prod
 
 echo "Building site with Hugo..."
 hugo # if using a theme, replace with `hugo -t <YOURTHEME>`
